+++ axelor-sale/src/main/java/com/axelor/apps/sale/service/saleorder/SaleOrderService.java
@@ -47,6 +47,14 @@ public interface SaleOrderService {
   void enableEditOrder(SaleOrder saleOrder) throws AxelorException;
 
   /**
+   * Check modified confirmed order before saving it.
+   * @param saleOrder
+   * @param saleOrderView
+   * @throws AxelorException
+   */
+  void checkModifiedConfirmedOrder(SaleOrder saleOrder, SaleOrder saleOrderView) throws AxelorException;
+
+  /**
    * Validate changes.
    *
    * @param saleOrder