+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - PARTNER BALANCE REPORT : edition date is always empty.
 - Shipping Coef: company is now required.
 - Purchase Order Controller: Better exception management.
+- Sale Order: added missing check on confirmed sale order modification
 
 
 ## [5.1.3] - 2019-02-21